Minor changes.
Reworked API endpoint probing to be a *bit* better. 3rd party JSON engine reduces up to 2 WHOLE seconds... off a 2 minute operation... Proper use of the JSON library instead of io.ReadAll saves a whole additional second and a half for a total or 3.5 seconds. worth it? Probably not, but I'll pretend for my sanity.

spammed concurrency for more "efficient" probing of the API at app startup, not that it matters though.

Finally started development on the chart portion of this bot. eta son